requirements.txt
beautifulsoup4==4.6.0
selenium==4.9.0
sendgrid==6.4.6
toml==0.10.1
ConfigArgParse==0.11.0
Pillow==9.5.0